test: split namespace environment white-box suites
