+++ crates/agent-engine/src/runtime/agent_loop/namespace_environment.rs
@@ -2420,1328 +2420,4 @@ fn agent_pid_from_path(agent_path: &str) -> Result<&str> {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::sync::{
-        Arc,
-        atomic::{AtomicUsize, Ordering as AtomicOrdering},
-    };
-
-    use alan_agentfs::{AgentConformanceChecker, AgentFs, AgentRootFs};
-    use alan_ap::{FileServer, InProcessTransport, Qid, Request};
-    use alan_kernel::{
-        Access, Credentials, MountFs, Namespace, ProcFs, ProcessInvocation, ProcessOutcome,
-        ProcessRunner,
-    };
-    use alan_llm::{
-        GenerationRequest, GenerationResponse, LlmProvider, MockLlmProvider, StreamChunk,
-    };
-    use alan_llmfs::LlmFs;
-    use alan_shell::Shell;
-    use tokio::sync::Notify;
-
-    use super::*;
-
-    struct EchoRunner;
-
-    #[async_trait::async_trait]
-    impl ProcessRunner for EchoRunner {
-        async fn run(&self, invocation: ProcessInvocation) -> ProcessOutcome {
-            let Ok(resolved) = invocation.namespace.resolve(&invocation.exec.executable) else {
-                return ProcessOutcome::exited(127, b"executable is not mounted\n".to_vec());
-            };
-            let fid = Fid(60_000 + invocation.pid.0);
-            let reachable = resolved
-                .call(Request::Walk {
-                    fid: Fid::ROOT,
-                    newfid: fid,
-                    names: resolved.rel.clone(),
-                })
-                .await
-                .is_ok();
-            let _ = resolved.call(Request::Clunk { fid }).await;
-            if !reachable {
-                return ProcessOutcome::exited(127, b"executable is not reachable\n".to_vec());
-            }
-            let mut output = invocation.exec.args.join(" ").into_bytes();
-            output.push(b'\n');
-            ProcessOutcome::exited(0, output)
-        }
-    }
-
-    struct LargeOutputRunner;
-
-    #[async_trait::async_trait]
-    impl ProcessRunner for LargeOutputRunner {
-        async fn run(&self, _invocation: ProcessInvocation) -> ProcessOutcome {
-            ProcessOutcome::exited(0, vec![b'x'; 70 * 1024])
-        }
-    }
-
-    struct ExactChunkOutputRunner;
-
-    #[async_trait::async_trait]
-    impl ProcessRunner for ExactChunkOutputRunner {
-        async fn run(&self, _invocation: ProcessInvocation) -> ProcessOutcome {
-            ProcessOutcome::exited(0, vec![b'y'; 64 * 1024])
-        }
-    }
-
-    struct LogicalFailureRunner;
-
-    #[async_trait::async_trait]
-    impl ProcessRunner for LogicalFailureRunner {
-        async fn run(&self, _invocation: ProcessInvocation) -> ProcessOutcome {
-            ProcessOutcome::exited(
-                0,
-                b"{\"success\":false,\"exit_code\":2,\"error\":\"command failed\"}\n".to_vec(),
-            )
-        }
-    }
-
-    struct AbortObservedRunner {
-        started: Arc<Notify>,
-        dropped: Arc<Notify>,
-    }
-
-    struct AbortDropGuard {
-        dropped: Arc<Notify>,
-    }
-
-    impl Drop for AbortDropGuard {
-        fn drop(&mut self) {
-            self.dropped.notify_one();
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl ProcessRunner for AbortObservedRunner {
-        async fn run(&self, _invocation: ProcessInvocation) -> ProcessOutcome {
-            let _guard = AbortDropGuard {
-                dropped: Arc::clone(&self.dropped),
-            };
-            self.started.notify_one();
-            std::future::pending::<ProcessOutcome>().await
-        }
-    }
-
-    struct BlockingStreamProvider {
-        started: Arc<Notify>,
-    }
-
-    #[async_trait::async_trait]
-    impl LlmProvider for BlockingStreamProvider {
-        async fn generate(
-            &mut self,
-            _request: GenerationRequest,
-        ) -> anyhow::Result<GenerationResponse> {
-            Err(anyhow::anyhow!("blocking provider uses streaming"))
-        }
-
-        async fn chat(&mut self, _system: Option<&str>, _user: &str) -> anyhow::Result<String> {
-            Ok("blocking stream provider".to_string())
-        }
-
-        async fn generate_stream(
-            &mut self,
-            _request: GenerationRequest,
-        ) -> anyhow::Result<tokio::sync::mpsc::Receiver<StreamChunk>> {
-            self.started.notify_one();
-            let (tx, rx) = tokio::sync::mpsc::channel(1);
-            tokio::spawn(async move {
-                let _hold = tx;
-                std::future::pending::<()>().await;
-            });
-            Ok(rx)
-        }
-
-        fn provider_name(&self) -> &'static str {
-            "blocking_stream"
-        }
-    }
-
-    fn tool_test_environment(
-        runner: Arc<dyn ProcessRunner>,
-    ) -> (NamespaceRuntimeEnvironment, Shell) {
-        let procfs = ProcFs::new().with_runner(runner);
-        let agentfs = Arc::new(AgentFs::new());
-        let binfs = Arc::new(alan_ap::reference::MemFs::new());
-
-        let mut child_namespace = Namespace::new();
-        child_namespace.mount(
-            "/proc",
-            InProcessTransport::new(Arc::new(procfs.clone())),
-            Access::ReadWrite,
-        );
-        child_namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs.clone()),
-            Access::ReadWrite,
-        );
-        child_namespace.mount("/bin", InProcessTransport::new(binfs), Access::ReadOnly);
-
-        let spawner_procfs =
-            Arc::new(procfs.for_spawner(None, child_namespace, Credentials::user("root-agent")));
-        let mut root_namespace = Namespace::new();
-        root_namespace.mount(
-            "/proc",
-            InProcessTransport::new(spawner_procfs),
-            Access::ReadWrite,
-        );
-        root_namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(root_namespace)));
-        (
-            NamespaceRuntimeEnvironment::new(root.clone(), "/agent/1", "default"),
-            Shell::new(root),
-        )
-    }
-
-    struct BlockingReadFs {
-        read_started: Notify,
-        clunked: Notify,
-        clunk_count: AtomicUsize,
-    }
-
-    impl BlockingReadFs {
-        fn new() -> Self {
-            Self {
-                read_started: Notify::new(),
-                clunked: Notify::new(),
-                clunk_count: AtomicUsize::new(0),
-            }
-        }
-
-        fn qid(kind: FileKind) -> Qid {
-            Qid {
-                kind,
-                version: 0,
-                path: 1,
-            }
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FileServer for BlockingReadFs {
-        async fn walk(
-            &self,
-            _fid: Fid,
-            _newfid: Fid,
-            _names: &[String],
-        ) -> std::result::Result<Qid, ErrorCode> {
-            Ok(Self::qid(FileKind::Stream))
-        }
-
-        async fn open(&self, _fid: Fid, _mode: OpenMode) -> std::result::Result<Qid, ErrorCode> {
-            Ok(Self::qid(FileKind::Stream))
-        }
-
-        async fn read(
-            &self,
-            _fid: Fid,
-            _offset: u64,
-            _count: u32,
-        ) -> std::result::Result<Vec<u8>, ErrorCode> {
-            self.read_started.notify_one();
-            std::future::pending().await
-        }
-
-        async fn write(
-            &self,
-            _fid: Fid,
-            _offset: u64,
-            _data: &[u8],
-        ) -> std::result::Result<u32, ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn stat(&self, _fid: Fid) -> std::result::Result<Stat, ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn create(
-            &self,
-            _fid: Fid,
-            _newfid: Fid,
-            _name: &str,
-            _kind: FileKind,
-        ) -> std::result::Result<Qid, ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn remove(&self, _fid: Fid) -> std::result::Result<(), ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn clunk(&self, _fid: Fid) -> std::result::Result<(), ErrorCode> {
-            self.clunk_count.fetch_add(1, AtomicOrdering::SeqCst);
-            self.clunked.notify_one();
-            Ok(())
-        }
-    }
-
-    struct ScriptedReadFs {
-        data: Vec<u8>,
-        kind: FileKind,
-        stat_length: Option<u64>,
-        clunk_count: AtomicUsize,
-    }
-
-    impl ScriptedReadFs {
-        fn new(data: impl Into<Vec<u8>>) -> Self {
-            Self {
-                data: data.into(),
-                kind: FileKind::Stream,
-                stat_length: None,
-                clunk_count: AtomicUsize::new(0),
-            }
-        }
-
-        fn shrinking_file(data: impl Into<Vec<u8>>, reported_length: u64) -> Self {
-            Self {
-                data: data.into(),
-                kind: FileKind::File,
-                stat_length: Some(reported_length),
-                clunk_count: AtomicUsize::new(0),
-            }
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FileServer for ScriptedReadFs {
-        async fn walk(
-            &self,
-            _fid: Fid,
-            _newfid: Fid,
-            _names: &[String],
-        ) -> std::result::Result<Qid, ErrorCode> {
-            Ok(BlockingReadFs::qid(self.kind))
-        }
-
-        async fn open(&self, _fid: Fid, _mode: OpenMode) -> std::result::Result<Qid, ErrorCode> {
-            Ok(BlockingReadFs::qid(self.kind))
-        }
-
-        async fn read(
-            &self,
-            _fid: Fid,
-            offset: u64,
-            count: u32,
-        ) -> std::result::Result<Vec<u8>, ErrorCode> {
-            let start = offset as usize;
-            if start >= self.data.len() {
-                return Ok(Vec::new());
-            }
-            let end = (start + count as usize).min(self.data.len());
-            Ok(self.data[start..end].to_vec())
-        }
-
-        async fn write(
-            &self,
-            _fid: Fid,
-            _offset: u64,
-            _data: &[u8],
-        ) -> std::result::Result<u32, ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn stat(&self, _fid: Fid) -> std::result::Result<Stat, ErrorCode> {
-            self.stat_length
-                .map(|length| Stat {
-                    name: "status".to_string(),
-                    qid: BlockingReadFs::qid(self.kind),
-                    length,
-                    executable: false,
-                    writable: false,
-                })
-                .ok_or(ErrorCode::Unsupported)
-        }
-
-        async fn create(
-            &self,
-            _fid: Fid,
-            _newfid: Fid,
-            _name: &str,
-            _kind: FileKind,
-        ) -> std::result::Result<Qid, ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn remove(&self, _fid: Fid) -> std::result::Result<(), ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn clunk(&self, _fid: Fid) -> std::result::Result<(), ErrorCode> {
-            self.clunk_count.fetch_add(1, AtomicOrdering::SeqCst);
-            Ok(())
-        }
-    }
-
-    #[tokio::test]
-    async fn read_file_accepts_regular_file_shrinking_after_stat() {
-        let fs = Arc::new(ScriptedReadFs::shrinking_file(b"exited\n", 8));
-        let client = NamespaceClient::new(InProcessTransport::new(fs.clone()));
-
-        assert_eq!(
-            client.read_file("/proc/10/status").await.unwrap(),
-            b"exited\n"
-        );
-        assert_eq!(fs.clunk_count.load(AtomicOrdering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn read_file_stops_at_reported_regular_file_length() {
-        let fs = Arc::new(ScriptedReadFs::shrinking_file(b"ready\nignored", 6));
-        let client = NamespaceClient::new(InProcessTransport::new(fs.clone()));
-
-        assert_eq!(
-            client.read_file("/proc/10/status").await.unwrap(),
-            b"ready\n"
-        );
-        assert_eq!(fs.clunk_count.load(AtomicOrdering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn read_stream_clunks_open_fid_when_cancelled() {
-        let fs = Arc::new(BlockingReadFs::new());
-        let client = NamespaceClient::new(InProcessTransport::new(fs.clone()));
-        let task = tokio::spawn({
-            let client = client.clone();
-            async move { client.read_stream_from("/agent/1/io/input", 0).await }
-        });
-
-        tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            fs.read_started.notified(),
-        )
-        .await
-        .expect("read should start");
-        task.abort();
-        let _ = task.await;
-        tokio::time::timeout(std::time::Duration::from_secs(1), fs.clunked.notified())
-            .await
-            .expect("cancelled read should clunk the fid");
-
-        assert_eq!(fs.clunk_count.load(AtomicOrdering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn llmfs_event_error_clunks_events_fid() {
-        let fs = Arc::new(ScriptedReadFs::new(
-            b"{\"version\":1,\"error\":\"temporary 503\"}\n".as_slice(),
-        ));
-        let client = NamespaceClient::new(InProcessTransport::new(fs.clone()));
-        let mut ignore = |_event: Event| async {};
-
-        let err =
-            read_generation_response_with_text_events(&client, "default", "gen-1", &mut ignore)
-                .await
-                .unwrap_err();
-
-        assert!(
-            err.to_string().contains("llmfs generation failed"),
-            "{err:#}"
-        );
-        assert_eq!(fs.clunk_count.load(AtomicOrdering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn controlled_generation_aborts_llmfs_on_cancel() {
-        let started = Arc::new(Notify::new());
-        let llmfs = Arc::new(LlmFs::new());
-        llmfs.register_connection(
-            "default",
-            Box::new(BlockingStreamProvider {
-                started: Arc::clone(&started),
-            }),
-        );
-        let agentfs = Arc::new(AgentFs::new());
-        let mut ns = Namespace::new();
-        ns.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        ns.mount(
-            "/mnt/llm",
-            InProcessTransport::new(llmfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-        let cancel = CancellationToken::new();
-        let task = tokio::spawn({
-            let environment = environment.clone();
-            let cancel = cancel.clone();
-            async move {
-                let request = GenerationRequest::new().with_user_message("hello");
-                let mut ignore = |_event: Event| async {};
-                environment
-                    .generate_with_text_events_controlled(&request, &mut ignore, 30, &cancel)
-                    .await
-            }
-        });
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), started.notified())
-            .await
-            .expect("provider stream should start");
-        cancel.cancel();
-        let err = task.await.unwrap().unwrap_err();
-        let err = format!("{err:#}");
-        assert!(
-            err.contains("cancelled") || err.contains("aborted"),
-            "{err}"
-        );
-
-        let status = String::from_utf8(
-            shell
-                .cat("/mnt/llm/connections/default/g0/status")
-                .await
-                .unwrap(),
-        )
-        .unwrap();
-        let status: serde_json::Value = serde_json::from_str(&status).unwrap();
-        assert_eq!(status["status"], "aborted");
-    }
-
-    #[tokio::test]
-    async fn m2_shell_talks_to_agent_through_files() {
-        let procfs = Arc::new(ProcFs::new());
-        let proc_server: Arc<dyn FileServer> = procfs.clone();
-        let agentfs = Arc::new(AgentFs::new());
-        let agent_root = Arc::new(AgentRootFs::new(proc_server));
-        let llmfs = Arc::new(LlmFs::new());
-        llmfs.register_connection(
-            "default",
-            Box::new(MockLlmProvider::new().with_response(GenerationResponse {
-                content: "hello from llmfs".to_string(),
-                thinking: None,
-                thinking_signature: None,
-                redacted_thinking: Vec::new(),
-                tool_calls: Vec::new(),
-                usage: None,
-                finish_reason: None,
-                provider_response_id: None,
-                provider_response_status: None,
-                warnings: Vec::new(),
-            })),
-        );
-
-        let mut ns = Namespace::new();
-        ns.mount("/proc", InProcessTransport::new(procfs), Access::ReadWrite);
-        ns.mount(
-            "/agent",
-            InProcessTransport::new(agent_root.clone()),
-            Access::ReadWrite,
-        );
-        ns.mount(
-            "/mnt/llm",
-            InProcessTransport::new(llmfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-
-        let pid = shell
-            .spawn(r#"{"executable":"/bin/agent","args":[]}"#)
-            .await
-            .unwrap();
-        assert_eq!(pid, "1");
-        agent_root.bind_process(pid.clone(), agentfs.clone()).await;
-        agent_root.set_root_process(pid.clone()).await;
-
-        shell
-            .write("/agent/1/io/input", b"hello agent")
-            .await
-            .unwrap();
-        let mut output_tail = shell.tail("/agent/1/io/output").await.unwrap();
-
-        let mut runtime = NamespaceTurnRuntime::new(
-            root.clone(),
-            NamespaceTurnRuntimeConfig::new("/agent/1", "default")
-                .with_system_prompt("You are an M2 test agent."),
-        );
-        let turn = runtime.run_next_turn().await.unwrap();
-
-        assert_eq!(turn.input, "hello agent");
-        assert_eq!(turn.response, "hello from llmfs");
-        assert!(!turn.generation_id.is_empty());
-
-        let streamed = output_tail.read(64 * 1024).await.unwrap();
-        output_tail.close().await.unwrap();
-        assert_eq!(String::from_utf8(streamed).unwrap(), "hello from llmfs");
-
-        let tape = String::from_utf8(shell.cat("/agent/1/machine/tape").await.unwrap()).unwrap();
-        assert!(tape.contains(r#""role":"user""#), "{tape}");
-        assert!(tape.contains(r#""content":"hello agent""#), "{tape}");
-        assert!(tape.contains(r#""role":"assistant""#), "{tape}");
-        assert!(tape.contains(r#""content":"hello from llmfs""#), "{tape}");
-        let checkpoint = runtime.current_tape_checkpoint().await.unwrap();
-        let checkpoint_file = String::from_utf8(
-            shell
-                .cat("/agent/1/machine/checkpoints/current")
-                .await
-                .unwrap(),
-        )
-        .unwrap();
-        assert_eq!(checkpoint, checkpoint_file.trim());
-        assert!(checkpoint.starts_with("sha256:"), "{checkpoint}");
-
-        AgentConformanceChecker::new(root)
-            .check_agent_process("/agent/1")
-            .await
-            .assert_ok();
-    }
-
-    #[tokio::test]
-    async fn engine_writes_requests_and_actions_as_agent_files() {
-        let procfs = Arc::new(ProcFs::new());
-        let proc_server: Arc<dyn FileServer> = procfs.clone();
-        let agentfs = Arc::new(AgentFs::new());
-        let agent_root = Arc::new(AgentRootFs::new(proc_server));
-        let mut ns = Namespace::new();
-        ns.mount("/proc", InProcessTransport::new(procfs), Access::ReadWrite);
-        ns.mount(
-            "/agent",
-            InProcessTransport::new(agent_root.clone()),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let pid = shell
-            .spawn(r#"{"executable":"/bin/agent","args":[]}"#)
-            .await
-            .unwrap();
-        assert_eq!(pid, "1");
-        agent_root.bind_process(pid.clone(), agentfs.clone()).await;
-        agent_root.set_root_process(pid).await;
-        let environment = NamespaceRuntimeEnvironment::new(root.clone(), "/agent/1", "default");
-
-        let request_id = environment
-            .write_request(
-                NamespaceRequestRecord::new("confirmation", "approve this action?")
-                    .with_options(r#"{"choices":["approve","deny"]}"#),
-            )
-            .await
-            .unwrap();
-        assert_eq!(request_id, "r0");
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/requests/{request_id}/kind"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "confirmation"
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/requests/{request_id}/prompt"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "approve this action?"
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/requests/{request_id}/options"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            r#"{"choices":["approve","deny"]}"#
-        );
-
-        let action_id = environment
-            .write_action(
-                NamespaceActionRecord::new("read", "completed")
-                    .with_output("file contents")
-                    .with_result(r#"{"ok":true}"#)
-                    .with_approval("not_required")
-                    .with_process("/proc/42"),
-            )
-            .await
-            .unwrap();
-        assert_eq!(action_id, "a0");
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/actions/{action_id}/name"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "read"
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/actions/{action_id}/status"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "completed"
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/actions/{action_id}/output"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "file contents"
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/actions/{action_id}/result"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            r#"{"ok":true}"#
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/actions/{action_id}/approval"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "not_required"
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/agent/1/actions/{action_id}/process"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "/proc/42"
-        );
-
-        AgentConformanceChecker::new(root)
-            .check_agent_process("/agent/1")
-            .await
-            .assert_ok();
-    }
-
-    #[tokio::test]
-    async fn answered_request_response_resumes_engine_pending_yield_from_files() {
-        let agentfs = Arc::new(AgentFs::new());
-        let mut ns = Namespace::new();
-        ns.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root.clone(), "/agent/1", "default");
-
-        let request_id = environment
-            .write_request(NamespaceRequestRecord::new(
-                "structured_input",
-                "Provide the missing detail",
-            ))
-            .await
-            .unwrap();
-        assert_eq!(request_id, "r0");
-        assert!(
-            environment
-                .resume_submission_from_answered_request(&request_id)
-                .await
-                .unwrap()
-                .is_none()
-        );
-
-        shell
-            .write(
-                &format!("/agent/1/requests/{request_id}/response"),
-                br#"{"answers":[{"question_id":"q1","value":"answer from request file"}]}"#,
-            )
-            .await
-            .unwrap();
-        let submission = environment
-            .resume_submission_from_answered_request(&request_id)
-            .await
-            .unwrap()
-            .expect("answered request becomes a resume submission");
-        match &submission.op {
-            Op::Resume {
-                request_id: resumed_id,
-                content,
-            } => {
-                assert_eq!(resumed_id, "r0");
-                assert_eq!(
-                    content,
-                    &vec![ContentPart::structured(serde_json::json!({
-                        "answers": [{"question_id": "q1", "value": "answer from request file"}]
-                    }))]
-                );
-            }
-            other => panic!("expected Op::Resume, got {other:?}"),
-        }
-
-        let mut turn_state = super::super::super::turn_state::TurnState::default();
-        turn_state.set_structured_input(crate::approval::PendingStructuredInputRequest {
-            request_id,
-            title: "Missing detail".to_string(),
-            prompt: "Provide the missing detail".to_string(),
-            questions: Vec::new(),
-        });
-        let mut state = super::super::RuntimeLoopState {
-            machine: crate::AgentMachine::new(),
-            current_submission_id: None,
-            environment,
-            core_config: crate::Config::default(),
-            runtime_config: super::super::super::RuntimeConfig::default(),
-            definition_persona_dirs: Vec::new(),
-            prompt_cache: super::super::super::prompt_cache::PromptAssemblyCache::new(Vec::new()),
-            turn_state,
-        };
-        let cancel = tokio_util::sync::CancellationToken::new();
-        let mut events = Vec::new();
-        let mut emit = |event| {
-            events.push(event);
-            async {}
-        };
-
-        let action = super::super::super::submission_handlers::handle_runtime_op_with_cancel(
-            &mut state,
-            submission.op,
-            &mut emit,
-            &cancel,
-        )
-        .await
-        .unwrap();
-
-        assert!(
-            matches!(
-                action,
-                super::super::super::submission_handlers::RuntimeOpAction::RunTurn { .. }
-            ),
-            "resume should re-enter the turn path: {action:?}"
-        );
-        assert!(!state.turn_state.has_pending_interaction());
-        assert!(events.is_empty());
-        let messages = state.machine.tape.messages();
-        assert_eq!(messages.len(), 1);
-        assert_eq!(messages[0].tool_responses()[0].id, "r0");
-        assert_eq!(
-            messages[0].tool_responses()[0].text_content(),
-            r#"{"answers":[{"question_id":"q1","value":"answer from request file"}]}"#
-        );
-    }
-
-    #[tokio::test]
-    async fn pending_request_selection_ignores_lexicographic_id_order() {
-        let agentfs = Arc::new(AgentFs::new());
-        let mut ns = Namespace::new();
-        ns.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        for index in 0..11 {
-            let id = environment
-                .write_request(NamespaceRequestRecord::new("confirmation", "approve?"))
-                .await
-                .unwrap();
-            assert_eq!(id, format!("r{index}"));
-            if index < 10 {
-                shell
-                    .write(&format!("/agent/1/requests/{id}/response"), b"approved")
-                    .await
-                    .unwrap();
-            }
-        }
-
-        let ids = environment.request_ids().await.unwrap();
-        assert_eq!(
-            environment.pending_request_id(&ids).await.unwrap(),
-            Some("r10".into())
-        );
-    }
-
-    #[tokio::test]
-    async fn machine_ctl_records_become_control_submissions_in_order() {
-        let agentfs = Arc::new(AgentFs::new());
-        let mut ns = Namespace::new();
-        ns.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        shell
-            .write("/agent/1/io/output", b"assistant output")
-            .await
-            .unwrap();
-        shell
-            .write("/agent/1/machine/ctl", b"compact")
-            .await
-            .unwrap();
-        shell
-            .write("/agent/1/machine/ctl", b"rollback")
-            .await
-            .unwrap();
-
-        let compact = environment
-            .read_next_machine_control_submission()
-            .await
-            .unwrap()
-            .expect("compact command should produce a submission");
-        assert!(matches!(compact.op, Op::CompactWithOptions { focus: None }));
-
-        let rollback = environment
-            .read_next_machine_control_submission()
-            .await
-            .unwrap()
-            .expect("rollback command should produce a submission");
-        assert!(matches!(rollback.op, Op::Rollback { turns: 1 }));
-
-        assert!(
-            environment
-                .read_next_machine_control_submission()
-                .await
-                .unwrap()
-                .is_none()
-        );
-
-        // Turn interrupt is a machine/ctl verb: a file client's Esc must
-        // cancel the running turn without touching kernel process lifecycle.
-        shell
-            .write("/agent/1/machine/ctl", b"interrupt")
-            .await
-            .unwrap();
-        let interrupt = environment
-            .read_next_machine_control_submission()
-            .await
-            .unwrap()
-            .expect("interrupt command should produce a submission");
-        assert!(matches!(interrupt.op, Op::Interrupt));
-    }
-
-    #[tokio::test]
-    async fn input_frame_becomes_engine_input_submission() {
-        let agentfs = Arc::new(AgentFs::new());
-        let mut ns = Namespace::new();
-        ns.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        shell
-            .write("/agent/1/io/input", b"continue from files")
-            .await
-            .unwrap();
-        let submission = environment
-            .read_next_input_submission(InputMode::FollowUp)
-            .await
-            .unwrap();
-
-        match submission.op {
-            Op::Input { parts, mode } => {
-                assert_eq!(mode, InputMode::FollowUp);
-                assert_eq!(parts, vec![ContentPart::text("continue from files")]);
-            }
-            other => panic!("expected Op::Input, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn input_frame_larger_than_initial_read_becomes_submission() {
-        let agentfs = Arc::new(AgentFs::new());
-        let mut ns = Namespace::new();
-        ns.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-        let message = "x".repeat(70 * 1024);
-
-        shell
-            .write("/agent/1/io/input", message.as_bytes())
-            .await
-            .unwrap();
-        let submission = environment
-            .read_next_input_submission(InputMode::FollowUp)
-            .await
-            .unwrap();
-
-        match submission.op {
-            Op::Input { parts, mode } => {
-                assert_eq!(mode, InputMode::FollowUp);
-                assert_eq!(parts, vec![ContentPart::text(message)]);
-            }
-            other => panic!("expected Op::Input, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn engine_spawns_process_with_spawner_assembled_namespace() {
-        let procfs = ProcFs::new();
-        let agentfs = Arc::new(AgentFs::new());
-        let llmfs = Arc::new(LlmFs::new());
-        let binfs = Arc::new(alan_ap::reference::MemFs::new());
-
-        let mut child_namespace = Namespace::new();
-        child_namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        child_namespace.mount(
-            "/mnt/llm",
-            InProcessTransport::new(llmfs),
-            Access::ReadWrite,
-        );
-        child_namespace.mount("/bin", InProcessTransport::new(binfs), Access::ReadOnly);
-
-        let spawner_procfs =
-            Arc::new(procfs.for_spawner(None, child_namespace, Credentials::user("root-agent")));
-        let mut root_namespace = Namespace::new();
-        root_namespace.mount(
-            "/proc",
-            InProcessTransport::new(spawner_procfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(root_namespace)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/root", "default");
-
-        let pid = environment
-            .spawn_process("/bin/agent", Vec::<String>::new())
-            .await
-            .unwrap();
-
-        assert_eq!(pid, "1");
-        assert_eq!(
-            String::from_utf8(shell.cat(&format!("/proc/{pid}/status")).await.unwrap()).unwrap(),
-            "running\n"
-        );
-        assert_eq!(
-            String::from_utf8(
-                shell
-                    .cat(&format!("/proc/{pid}/credentials"))
-                    .await
-                    .unwrap()
-            )
-            .unwrap(),
-            "root-agent"
-        );
-        let namespace =
-            String::from_utf8(shell.cat(&format!("/proc/{pid}/namespace")).await.unwrap()).unwrap();
-        assert!(namespace.lines().any(|line| line == "/agent/1 rw"));
-        assert!(namespace.lines().any(|line| line == "/mnt/llm rw"));
-        assert!(namespace.lines().any(|line| line == "/bin ro"));
-    }
-
-    #[tokio::test]
-    async fn engine_runs_tool_as_process_and_projects_action_files() {
-        let procfs = ProcFs::new().with_runner(Arc::new(EchoRunner));
-        let agentfs = Arc::new(AgentFs::new());
-        let binfs = Arc::new(alan_ap::reference::MemFs::new());
-
-        let mut child_namespace = Namespace::new();
-        child_namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs.clone()),
-            Access::ReadWrite,
-        );
-        child_namespace.mount("/bin", InProcessTransport::new(binfs), Access::ReadOnly);
-
-        let spawner_procfs =
-            Arc::new(procfs.for_spawner(None, child_namespace, Credentials::user("root-agent")));
-        let mut root_namespace = Namespace::new();
-        root_namespace.mount(
-            "/proc",
-            InProcessTransport::new(spawner_procfs),
-            Access::ReadWrite,
-        );
-        root_namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(root_namespace)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        let action = environment
-            .run_tool_action(
-                "echo",
-                "/bin/greeting",
-                ["hello".to_string(), "from-process".to_string()],
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(action.pid, "1");
-        assert_eq!(action.action_id, "a0");
-        assert_eq!(action.output, "hello from-process\n");
-        assert_eq!(action.exit_code, 0);
-        assert_eq!(
-            environment.read_process_exit_code("1").await.unwrap(),
-            Some(0)
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/status").await.unwrap()).unwrap(),
-            "exited\n"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/io/output").await.unwrap()).unwrap(),
-            "hello from-process\n"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/name").await.unwrap()).unwrap(),
-            "echo"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/status").await.unwrap()).unwrap(),
-            "completed"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/output").await.unwrap()).unwrap(),
-            "hello from-process\n"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/result").await.unwrap()).unwrap(),
-            r#"{"exit_code":0}"#
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/process").await.unwrap()).unwrap(),
-            "/proc/1"
-        );
-    }
-
-    #[tokio::test]
-    async fn run_tool_action_projects_logical_payload_failure_as_failed_action() {
-        let (environment, shell) = tool_test_environment(Arc::new(LogicalFailureRunner));
-
-        let action = environment
-            .run_tool_action("bash", "/bin/bash", ["{}"])
-            .await
-            .unwrap();
-
-        assert_eq!(action.pid, "1");
-        assert_eq!(action.action_id, "a0");
-        assert_eq!(action.exit_code, 2);
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/exit").await.unwrap()).unwrap(),
-            "0"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/status").await.unwrap()).unwrap(),
-            "failed"
-        );
-        let result =
-            String::from_utf8(shell.cat("/agent/1/actions/a0/result").await.unwrap()).unwrap();
-        let result: serde_json::Value = serde_json::from_str(&result).unwrap();
-        assert_eq!(result["exit_code"], 2);
-        assert_eq!(result["process_exit_code"], 0);
-    }
-
-    #[tokio::test]
-    async fn run_tool_action_cancels_spawned_process_on_cancel() {
-        let started = Arc::new(Notify::new());
-        let dropped = Arc::new(Notify::new());
-        let (environment, shell) = tool_test_environment(Arc::new(AbortObservedRunner {
-            started: Arc::clone(&started),
-            dropped: Arc::clone(&dropped),
-        }));
-        let cancel = CancellationToken::new();
-        let task = tokio::spawn({
-            let environment = environment.clone();
-            let cancel = cancel.clone();
-            async move {
-                environment
-                    .run_tool_action_with_cancel(
-                        "blocked",
-                        "/bin/blocked",
-                        Vec::<String>::new(),
-                        &cancel,
-                    )
-                    .await
-            }
-        });
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), started.notified())
-            .await
-            .expect("tool runner should start");
-        cancel.cancel();
-        let err = task.await.unwrap().unwrap_err();
-        assert!(err.to_string().contains("cancelled"), "{err:#}");
-        tokio::time::timeout(std::time::Duration::from_secs(1), dropped.notified())
-            .await
-            .expect("tool runner future should be aborted");
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/status").await.unwrap()).unwrap(),
-            "exited\n"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/exit").await.unwrap()).unwrap(),
-            "130"
-        );
-    }
-
-    #[tokio::test]
-    async fn run_tool_action_cancels_spawned_process_on_wait_timeout() {
-        let started = Arc::new(Notify::new());
-        let dropped = Arc::new(Notify::new());
-        let (environment, shell) = tool_test_environment(Arc::new(AbortObservedRunner {
-            started: Arc::clone(&started),
-            dropped: Arc::clone(&dropped),
-        }));
-        let cancel = CancellationToken::new();
-        let task = tokio::spawn({
-            let environment = environment.clone();
-            let cancel = cancel.clone();
-            async move {
-                environment
-                    .run_tool_action_with_cancel_and_timeout(
-                        "blocked",
-                        "/bin/blocked",
-                        Vec::<String>::new(),
-                        &cancel,
-                        1,
-                    )
-                    .await
-            }
-        });
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), started.notified())
-            .await
-            .expect("tool runner should start");
-        let err = tokio::time::timeout(std::time::Duration::from_secs(2), task)
-            .await
-            .expect("tool wait should use the configured timeout")
-            .unwrap()
-            .unwrap_err();
-        let err = format!("{err:#}");
-        assert!(err.contains("timed out waiting 1s"), "{err}");
-        tokio::time::timeout(std::time::Duration::from_secs(1), dropped.notified())
-            .await
-            .expect("tool runner future should be aborted on wait timeout");
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/status").await.unwrap()).unwrap(),
-            "exited\n"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/exit").await.unwrap()).unwrap(),
-            "130"
-        );
-    }
-
-    #[tokio::test]
-    async fn run_tool_action_reads_output_larger_than_initial_read() {
-        let (environment, _shell) = tool_test_environment(Arc::new(LargeOutputRunner));
-
-        let action = environment
-            .run_tool_action("large", "/bin/large", Vec::<String>::new())
-            .await
-            .unwrap();
-
-        assert_eq!(action.output.len(), 70 * 1024);
-        assert!(action.output.bytes().all(|byte| byte == b'x'));
-        assert_eq!(action.exit_code, 0);
-    }
-
-    #[tokio::test]
-    async fn run_tool_action_reads_exact_chunk_output_without_waiting_for_more() {
-        let (environment, _shell) = tool_test_environment(Arc::new(ExactChunkOutputRunner));
-
-        let action = tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            environment.run_tool_action("exact", "/bin/exact", Vec::<String>::new()),
-        )
-        .await
-        .expect("exact chunk output should not wait at the live stream edge")
-        .unwrap();
-
-        assert_eq!(action.output.len(), 64 * 1024);
-        assert!(action.output.bytes().all(|byte| byte == b'y'));
-        assert_eq!(action.exit_code, 0);
-    }
-
-    #[tokio::test]
-    async fn engine_tape_writer_holds_generating_lease_and_allows_readers() {
-        let agentfs = Arc::new(AgentFs::new());
-        let mut ns = Namespace::new();
-        ns.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(ns)));
-        let shell = Shell::new(root.clone());
-        let environment = NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        let mut writer = environment.begin_tape_generation().await.unwrap();
-
-        let second_writer = environment.begin_tape_generation().await;
-        assert!(
-            second_writer.is_err(),
-            "a second engine writer must not acquire machine/tape while GENERATING lease is held"
-        );
-
-        let mut tape_tail = shell.tail("/agent/1/machine/tape").await.unwrap();
-        writer.append_record("user", "hello").await.unwrap();
-        let streamed = String::from_utf8(tape_tail.read(64 * 1024).await.unwrap()).unwrap();
-        assert!(streamed.contains(r#""role":"user""#), "{streamed}");
-        assert!(streamed.contains(r#""content":"hello""#), "{streamed}");
-        tape_tail.close().await.unwrap();
-
-        writer.append_record("assistant", "hi").await.unwrap();
-        writer.finish().await.unwrap();
-
-        let mut next_writer = environment.begin_tape_generation().await.unwrap();
-        next_writer
-            .append_record("assistant", "after lease")
-            .await
-            .unwrap();
-        next_writer.finish().await.unwrap();
-
-        let tape = String::from_utf8(shell.cat("/agent/1/machine/tape").await.unwrap()).unwrap();
-        assert!(tape.contains(r#""content":"hi""#), "{tape}");
-        assert!(tape.contains(r#""content":"after lease""#), "{tape}");
-    }
-
-    #[test]
-    fn tape_record_shape_is_content_addressable_ready() {
-        let record = tape_record_bytes("assistant", "stable text").unwrap();
-        assert_eq!(
-            String::from_utf8(record).unwrap(),
-            r#"{"version":1,"kind":"message","role":"assistant","content":"stable text"}"#
-                .to_string()
-                + "\n",
-            "tape records must stay canonical, self-contained newline-delimited units"
-        );
-    }
-}
+mod tests;
